test: relocate the memory and duplication cases to their own files
Six tests move by location only. Each new file still constructs a real `AgentService`
through agent-service-test-harness.ts and asserts through public events, `listMemories`
/ `listBots`, and the JSONL protocol log — no assertion was rewritten, and none became a
unit test of the controller.

agent/agent-memories.test.ts takes the two staging cases from the questions shard:
commit-after-success with the refreshed next-turn context, and the discard case that
walks failed, interrupted and cleared turns. Both drive `remember` over the wire as
`item/tool/call`, which is the contract that matters and is not visible from inside the
controller.

agent/duplication-gate.test.ts takes the four duplication cases from the providers shard:
the full copy, the active-work refusal, the commit-lock serialisation, and the source-
changed rollback.

What stayed, and why. The questions shard keeps both prompt cases — a question prompt is
attention, not memory. The providers shard keeps everything else it had; none of the
remaining cases touch either controller. The restart shard keeps "deletes idle bots and
refuses to orphan active work" even though `deleteBot` now calls `releaseForDelete`: the
behaviour it names is deletion refusing to orphan work, and the duplication release is
incidental to it.

Verified: providers, questions and all nine agent/*.test.ts, 48 tests, plus
`bun run typecheck:node`.
